+++ src/lib/helpers/app-image/get-current-image.ts
@@ -28,41 +28,65 @@ export function getMaxLength(previewAppImage: PreviewDataAppImage, onlineImages:
 }
 
 // Set index modulo max length.
-export function setImageIndex(previewAppImage: PreviewDataAppImage, onlineImages: OnlineImages[ArtworkType], index: number) {
+export function setImageIndex(previewAppImage: PreviewDataAppImage, onlineImages: OnlineImages[ArtworkType], imageIndex: number) {
   const lens = getMaxLength(previewAppImage, onlineImages);
-  previewAppImage.imageIndex = mod(index, lens.maxLength)
+  previewAppImage.imageIndex = mod(imageIndex, lens.maxLength)
 }
 
-export function getCurrentImage(previewAppImage: PreviewDataAppImage, onlineImages: OnlineImages[ArtworkType]) {
+export function getImageRanges(previewAppImage: PreviewDataAppImage, onlineImages: OnlineImages[ArtworkType]) {
+  const lens = getMaxLength(previewAppImage, onlineImages);
+  let ranges: {[k: string]: {start: number, end: number}} = {};
+  ranges['default'] = {start: 0, end: lens.singleLength + lens.multiLength + lens.onlineLength == 0 ? lens.defaultLength : 0}
+  let integrated=0;
+  for(let n=0; n < singleLocalProviders.length; n++) {
+    ranges[singleLocalProviders[n]] = {start: integrated, end: integrated + lens.singleLocalLengths[singleLocalProviders[n]]}
+    integrated = ranges[singleLocalProviders[n]].end;
+  }
+  for(let n=0; n < multiLocalProviders.length; n++) {
+    ranges[multiLocalProviders[n]] = {start: integrated, end: integrated + lens.multiLocalLengths[multiLocalProviders[n]]}
+    integrated = ranges[multiLocalProviders[n]].end;
+  }
+  for(let n=0; n < onlineProviders.length; n++) {
+    ranges[onlineProviders[n]] = {start: integrated, end: integrated + lens.onlineLengths[onlineProviders[n]]};
+    integrated = ranges[onlineProviders[n]].end;
+  }
+  return ranges;
+}
+
+export function getImage(previewAppImage: PreviewDataAppImage, onlineImages: OnlineImages[ArtworkType], imageIndex: number) {
   const lens = getMaxLength(previewAppImage, onlineImages)
   if (lens.maxLength !== 0) {
-    setImageIndex(previewAppImage, onlineImages, previewAppImage.imageIndex)
-    if(lens.maxLength == 1 && lens.defaultLength == 1) {
+    imageIndex = mod(imageIndex, lens.maxLength);
+    if(lens.singleLength+lens.multiLength+lens.onlineLength == 0) {
       return previewAppImage.default
     }
     let integrated=0;
     for(let n=0; n < singleLocalProviders.length; n++) {
       integrated += lens.singleLocalLengths[singleLocalProviders[n]]
-      if(previewAppImage.imageIndex < integrated) {
+      if(imageIndex < integrated) {
         return previewAppImage.singleProviders[singleLocalProviders[n]]
       }
     }
     let offset = integrated;
     for(let n=0; n < multiLocalProviders.length; n++) {
       integrated += lens.multiLocalLengths[multiLocalProviders[n]]
-      if(previewAppImage.imageIndex < integrated) {
-        return onlineImages[previewAppImage.imagePool].offline[multiLocalProviders[n]][previewAppImage.imageIndex - offset]
+      if(imageIndex < integrated) {
+        return onlineImages[previewAppImage.imagePool].offline[multiLocalProviders[n]][imageIndex - offset]
       }
       offset = integrated;
     }
     for(let n=0; n < onlineProviders.length; n++) {
       integrated += lens.onlineLengths[onlineProviders[n]];
-      if(previewAppImage.imageIndex < integrated) {
-        return onlineImages[previewAppImage.imagePool].online[onlineProviders[n]].content[previewAppImage.imageIndex - offset]
+      if(imageIndex < integrated) {
+        return onlineImages[previewAppImage.imagePool].online[onlineProviders[n]].content[imageIndex - offset]
       }
       offset=integrated;
     }
   } else {
     return undefined
   }
+}
+
+export function getCurrentImage(previewAppImage: PreviewDataAppImage, onlineImages: OnlineImages[ArtworkType]) {
+  return getImage(previewAppImage, onlineImages, previewAppImage.imageIndex)
 }